Sistemato git e gitignore

+++ app/src/main/java/com/matteopaterno/progettopwm/hotel/HotelData.kt
@@ -3,6 +3,6 @@ package com.matteopaterno.progettopwm.hotel
 data class HotelData(
     var image: Int,
     var nome: String,
-    var rating: Double,
+    var rating: Float,
     var posizione: String
 )
